pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4"
pydantic-settings = ">=2.12.0,<3.0.0"
sqlalchemy = ">=2.0.46,<3.0.0"
structlog = ">=25.5.0,<26.0.0"
orjson = ">=3.11.7,<4.0.0"
inflection = ">=0.5.1,<0.6.0"
python-magic = ">=0.4.27,<0.5.0"
pypdf2 = ">=3.0.1,<4.0.0"
python-docx = ">=1.1.2,<2.0.0"
openpyxl = ">=3.1.5,<4.0.0"
pandas = ">=2.2.0,<3.0.0"
patool = ">=3.0.2,<4.0.0"
python-magic-bin = ">=0.4.14,<0.5.0"
reportlab = "^4.4.10"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
mypy = ">=1.15.0"
ruff = ">=0.4.8"
pre-commit = "^4.5.1"
pytest-mock = "^3.15.1"


[tool.ruff]
src = ["app"]
target-version = "py311"
line-length = 79
exclude = ["__pycache__/", ".git/", "*.pyc"]
format.indent-style = "space"
format.skip-magic-trailing-comma = false
format.quote-style = "double"
preview = true

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "RUF"]
ignore = ["RUF001"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
pretty = true
show_error_codes = true
disable_error_code = ["import-untyped"]
exclude = [
    'build/',
    'dist/',
    '\.git',
]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--asyncio-mode=auto"
]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "slow: slow running tests"
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/__pycache__/*",
    "*/tests/*",
]
